Change structure of the routing
